// src/app/modelos/usuario-dto.ts
export interface UsuarioDto {

    idusuario?: number;
    variable1?: string;
    variable2?: string;
    variable3?: string;
    variable4?: string;
    variable5?: string;
    variable6?: string;
    variable7?: string;
    variable8?: string;
    variable9?: string;
    variable10?: string;
    variable11?: string;
    variable12?: string;
    variable13?: string;
    variable14?: string;
    variable15?: string;
    variable16?: string;
    evento?: string;





}